Arrange FilterNotNull tests
