+++ src/core-taggeds-optional/Optional.Tests/Optional.Extensions.FilterNotNull.Test/FilterNotNullOrThrow.cs
@@ -52,5 +52,107 @@ namespace PrimeFuncPack.Core.Tests
             var ex = Assert.Throws<SomeException>(() => _ = source.FilterNotNullOrThrow(() => resultException));
             Assert.AreSame(resultException, ex);
         }
+
+        [Test]
+        public void FilterNotNullOrThrow_SourceValueIsNotNullRefType_ExpectPresentNotNullable()
+        {
+            var sourceValue = PlusFifteenIdRefType;
+            var source = Optional<RefType?>.Present(sourceValue);
+
+            var actual = source.FilterNotNullOrThrow();
+            var expected = Optional<RefType>.Present(sourceValue);
+
+            Assert.AreEqual(expected, actual);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrow_SourceValueIsNullRefType_ExpectInvalidOperationException()
+        {
+            var source = Optional<RefType?>.Present(null);
+            _ = Assert.Throws<InvalidOperationException>(() => _ = source.FilterNotNullOrThrow());
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_RefTypeExceptionFactoryIsNull_ExpectArgumentNullException()
+        {
+            var source = Optional<RefType?>.Present(PlusFifteenIdRefType);
+
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = source.FilterNotNullOrThrow(null!));
+            Assert.AreEqual("exceptionFactory", ex!.ParamName);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_SourceValueIsNotNullRefType_ExpectPresentNotNullable()
+        {
+            var sourceValue = PlusFifteenIdRefType;
+            var source = Optional<RefType?>.Present(sourceValue);
+
+            var actual = source.FilterNotNullOrThrow(() => new SomeException());
+            var expected = Optional<RefType>.Present(sourceValue);
+
+            Assert.AreEqual(expected, actual);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_SourceValueIsNullRefType_ExpectCreatedException()
+        {
+            var source = Optional<RefType?>.Present(null);
+            var createdException = new SomeException();
+
+            var actualExcepton = Assert.Throws<SomeException>(
+                () => _ = source.FilterNotNullOrThrow(() => createdException));
+            Assert.AreSame(createdException, actualExcepton);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrow_SourceValueIsNotNullStructType_ExpectPresentNotNullable()
+        {
+            var sourceValue = SomeTextStructType;
+            var source = Optional<StructType?>.Present(sourceValue);
+
+            var actual = source.FilterNotNullOrThrow();
+            var expected = Optional<StructType>.Present(sourceValue);
+
+            Assert.AreEqual(expected, actual);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrow_SourceValueIsNullStructType_ExpectInvalidOperationException()
+        {
+            var source = Optional<StructType?>.Present(null);
+            _ = Assert.Throws<InvalidOperationException>(() => _ = source.FilterNotNullOrThrow());
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_StructExceptionFactoryIsNull_ExpectArgumentNullException()
+        {
+            var source = Optional<StructType?>.Present(null);
+
+            var ex = Assert.Throws<ArgumentNullException>(() => _ = source.FilterNotNullOrThrow(null!));
+            Assert.AreEqual("exceptionFactory", ex!.ParamName);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_SourceValueIsNotNullStructType_ExpectPresentNotNullable()
+        {
+            var sourceValue = NullTextStructType;
+            var source = Optional<StructType?>.Present(sourceValue);
+
+            var actual = source.FilterNotNullOrThrow(() => new SomeException());
+            var expected = Optional<StructType>.Present(sourceValue);
+
+            Assert.AreEqual(expected, actual);
+        }
+
+        [Test]
+        public void FilterNotNullOrThrowWithFactory_SourceValueIsNullStructType_ExpectCreatedException()
+        {
+            var source = Optional<StructType?>.Present(null);
+            var createdException = new SomeException();
+
+            var actualExcepton = Assert.Throws<SomeException>(
+                () => _ = source.FilterNotNullOrThrow(() => createdException));
+            Assert.AreSame(createdException, actualExcepton);
+        }
     }
 }
